feat: í: 회원이름, 트레이너 이름으로 예약 검색 구현

+++ src/main/java/com/predev/gymcrm/dto/resp/AdminSearchTrainerRespDto.java
@@ -12,6 +12,6 @@ import lombok.NoArgsConstructor;
 public class AdminSearchTrainerRespDto {
     private int trainerId;
     private String name;
-    private int memberCount;
-    private float avgScore;
+    private Integer memberCount;
+    private Double avgScore;
 }
